+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+- [Feature] Added new Storybook examples detailing div styling helper classes ([#1391](https://github.com/optimizely/oui/pull/1391))
 
 ## 48.0.2 - 2020-08-19
 - [Patch] Update `flex--1` helper class ([#1405](https://github.com/optimizely/oui/pull/1405))